refactor app entry point
